Refactor naming conventions and improve PIE loading logic
Renamed classes and properties in `Exchange.cs` to follow consistent naming conventions, including changing `Workingschedule` to `WorkingSchedule` and `Timeevent` to `TimeEvent`. Updated the type of `WorkingSchedules` and `TimeEvents` properties for consistency.

In `Instrument.cs`, changed the type of `WorkingScheduleId` from `int` to `long`.

Refactored `SettingsForm.cs` to rename the PIE loading method from `LoadPIEs` to `GetPieDropdownItems`, enhancing error handling and adding checks for pie presence. Introduced a new method, `PopulatePieDropdown`, to manage dropdown population and added logic to disable the dropdown if only one PIE is available.

+++ StockDisplay/Domain/Instrument.cs
@@ -11,6 +11,6 @@
         public string ShortName { get; set; } = string.Empty;
         public string Ticker { get; set; } = string.Empty;
         public string Type { get; set; } = string.Empty;
-        public int WorkingScheduleId { get; set; }
+        public long WorkingScheduleId { get; set; }
     }
 }
